Add Beta prior - posterior changes
If you increase n, the posterior will ignore more and more the prior and base everything on the data.If n is small, the expected value tends towards the expected value of the prior. The expected value of the posterior is a mixture of the two expectations (prior and likelihood).

Signed-off-by: Erik-Cristian Seulean <erik@Erik-Cristians-MacBook-Air.local>